Refactor DruidSchema & DruidTable
* Pulls the segment data cache out of DruidSchema
* Breaks DruidTable into multiple classes
* More efficient table name lookup
