+++ core/src/test/java/org/apache/druid/data/input/impl/HttpInputSourceConfigTest.java
@@ -26,7 +26,6 @@ import org.junit.Test;
 
 public class HttpInputSourceConfigTest
 {
-
   @Test
   public void testEquals()
   {
