+++ sql/src/main/java/org/apache/druid/sql/calcite/external/ExternalTableScan.java
@@ -27,22 +27,24 @@ import org.apache.calcite.rel.AbstractRelNode;
 import org.apache.calcite.rel.RelWriter;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.druid.sql.calcite.table.DruidTable;
+import org.apache.druid.sql.calcite.table.ExternalTable;
 
 /**
- * Represents a scan of an external table. Generated by {@link DruidTable} when its datasource is an
- * {@link ExternalDataSource}.
+ * Represents a scan of an external table. Generated by {@link ExternalTable},
+ * which represents an {@link ExternalDataSource}.
  *
- * This class is exercised in CalciteInsertDmlTest but is not currently exposed to end users.
+ * This class is exercised in CalciteInsertDmlTest but is not currently
+ * exposed to end users.
  */
 public class ExternalTableScan extends AbstractRelNode
 {
   private final ObjectMapper jsonMapper;
-  private final DruidTable druidTable;
+  private final ExternalTable druidTable;
 
   public ExternalTableScan(
       final RelOptCluster cluster,
       final ObjectMapper jsonMapper,
-      final DruidTable druidTable
+      final ExternalTable druidTable
   )
   {
     super(cluster, cluster.traitSetOf(Convention.NONE));
