+++ sql/src/main/java/org/apache/druid/sql/calcite/schema/InformationSchema.java
@@ -59,6 +59,7 @@ import org.apache.druid.sql.calcite.table.DruidTable;
 import org.apache.druid.sql.calcite.table.RowSignatures;
 
 import javax.annotation.Nullable;
+
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
