+++ sql/src/test/java/org/apache/druid/sql/calcite/expression/ExpressionTestHelper.java
@@ -63,6 +63,7 @@ import org.joda.time.DateTimeZone;
 import org.junit.Assert;
 
 import javax.annotation.Nullable;
+
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Collections;
