+++ sql/src/test/java/org/apache/druid/sql/calcite/rule/DruidUnionDataSourceRuleTest.java
@@ -32,6 +32,8 @@ import org.apache.druid.sql.calcite.rel.DruidRel;
 import org.apache.druid.sql.calcite.rel.DruidRelsTest;
 import org.apache.druid.sql.calcite.rel.DruidUnionDataSourceRel;
 import org.apache.druid.sql.calcite.rel.PartialDruidQuery;
+import org.apache.druid.sql.calcite.table.DatasourceTable;
+import org.apache.druid.sql.calcite.table.DatasourceTable.DatasourceMetadata;
 import org.apache.druid.sql.calcite.table.DruidTable;
 import org.easymock.EasyMock;
 import org.junit.Assert;
@@ -43,16 +45,17 @@ import java.util.Optional;
 
 public class DruidUnionDataSourceRuleTest
 {
-  private final DruidTable fooDruidTable = new DruidTable(
-      new TableDataSource("foo"),
-      RowSignature.builder()
-                  .addTimeColumn()
-                  .add("col1", ColumnType.STRING)
-                  .add("col2", ColumnType.LONG)
-                  .build(),
-      null,
-      false,
-      false
+  private final DruidTable fooDruidTable = new DatasourceTable(
+      new DatasourceMetadata(
+          new TableDataSource("foo"),
+          RowSignature.builder()
+                      .addTimeColumn()
+                      .add("col1", ColumnType.STRING)
+                      .add("col2", ColumnType.LONG)
+                      .build(),
+          false,
+          false
+      )
   );
 
   @Test
